requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
mysqlclient==2.2.0
pymysql==1.1.0
pydantic[email]==2.4.2
emails==0.6
secure-smtplib==0.1.1